configure: if not explicitly enabled, disable frontends with missing deps
If the user does not explicitly enable a frontend, and the pre-requisites
for this frontned are missing, then disable this frontend.

By default, all frontends are enabled. But if (for example) the ncurses
library is missing, and the user did not explicitly enabled them, then
do not build these frontends.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>