Makefile.am-s: allow building outside of the source tree
Hi,

I saw your announcement about kconfig-frontends to uclibc@uclibc.org
the other day, and thought I would try it out for our own project
where we use an old derivative of the kconfig system.

However, to make it build with our build system I needed to apply
the patch below to allow building outside the source tree.

After that all worked fine, and the generated .config file is
still the same as with the old config system. :)

[--SNIP unrelated message--]

Signed-off-by: Peter Kjellerstedt <pkj@axis.com>