version: Redo the version string computation
In case the kconfig-frontends version is 'hg', always include
the Linux kernel changeset we last synced up with, and include
the potential '-rcX' if present.

In case the kconfig-frontends version is not 'hg', we assume
this is a digit-version, and we do not include hte kernel cset
in the version, but we keep an potential '-rcX' if present.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>