configure: avoid warning with automake 1.12 on Cygwin
With automake 1.12 came a new macro, AM_PROG_AR, which *must* be used if
the archiver (eg. ar on GNU systems) is used, to support exotic archivers
such as the one from Microsoft (lib).

Unfortunately, this macro causes issues on older automake versions, and
thus needs to be conditionalised (if that wasn't a word, it now is ;-) )

Furthermore, this macro *must* be called before we initialise libtool
(with LT_INIT).

Sigh... :-/

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>