scripts: use a dot to separate third and fourth digits
Traditionally, Linux distributions have been using the dash as a separator
between the package's upstream versionning and their own local packaging
versioning.

Using a dash in kconfig-frontends' own versioning scheme may confuse the
distributions, which may not expect this situation (and even if some do,
not all do).

In the beginings, I thought about using a dash to make it explicit that
the digits before the dash was Linux' version, while the digit(s) after
the dash was kconfig-frontends' version. That, because the Linux version
has only two digits for .0 releases, and three digits for stable releases.
But since the Linux kernel's Makefile still report a three-digit version
even for .0 releases, we ended up always using a three-digit Linux version,
and thus always had a four-digit kconfig-frontends' version.

So, use a dot instead of a dash.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>